[infra] Prevent front end unit tests from being approved.
This change adds a disguised test.py which the testing infrastructure does
not recognize, effectively ensuring the test results from the front end
unit tests are not uploaded to the database, preventing them from being
approved. As such those tests will be invisible to the testing
infrastructure and will not show up on dashboards.

Change-Id: Ia28976e7cd94284eee968b215a0023a6db52d9c1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/94753
Reviewed-by: William Hesse <whesse@google.com>