Issue 36890. Improve importing elements when incomplete code.
There might be some false negatives, when there *is* another top-level
declaration with the same name as we are removing. But this is a
pre-existing conflict, so I think we should not be very concerned. At
some point we might collect initial potentially conflicting names using
the same SyntacticScopeNamesCollector instead of element model. With
element model we don't know about parser recovery.

R=brianwilkerson@google.com

Bug: https://github.com/dart-lang/sdk/issues/36890
Change-Id: Ia05db299990f41f16feb493c2812237be84f3d2e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/101902
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>