[compiler] Fix EmitTestAndCall for a complete call of a Smi method.
In rare circumstances we can generate a complete monomorphic
PolymorphicInstanceCall targeting a method of Smi class.

In this case we don't need to check anything - we can simply emit a
static call.

(This fixes a bug in the incorrect fix submitted earlier as
c67b9f23316901ecc1fc24189c2837430eabdb33 - even if call is complete
we still need to check for a smi receiver if there are more possible
receiver classes)

Fixes #37019

Bug: 37019
Change-Id: If0250ba2811f00cc4fe595bf7cd03a7c0d820076
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/103809
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>
Commit-Queue: Vyacheslav Egorov <vegorov@google.com>