[vm] Ensure C-struct is initialized to guarantee nullptr in Dart_KernelCompilationResult.kernel
A number of places in the code will free `Dart_KernelCompilationResult.kernel`
if non-nullptr even if `Dart_KernelCompilationResult.error` is non-nullptr.

Normally the two are mutually exclusive (either `error` is set or `kernel` is set) but
some code frees both if they are non-null.

This issue has caused free() of uninitialized memory which in return
caused some of our benchmarks to break.

Change-Id: Ie65a0c51a32a3616361f3852c13c859e560285b5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/104563
Reviewed-by: Daco Harkes <dacoharkes@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>